+++ db.tf
@@ -4,7 +4,23 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  */
 
+data "vault_generic_secret" "apiarydb_master_user" {
+  count = "${ var.external_database_host == "" ? 1 : 0 }"
+  path  = "${local.vault_path}/db_master_user"
+}
+
+data "vault_generic_secret" "hive_rwuser" {
+  count = "${ var.external_database_host == "" ? 1 : 0 }"
+  path  = "${local.vault_path}/hive_rwuser"
+}
+
+data "vault_generic_secret" "hive_rouser" {
+  count = "${ var.external_database_host == "" ? 1 : 0 }"
+  path  = "${local.vault_path}/hive_rouser"
+}
+
 resource "aws_db_subnet_group" "apiarydbsg" {
+  count       = "${ var.external_database_host == "" ? 1 : 0 }"
   name        = "${local.instance_alias}-dbsg"
   subnet_ids  = ["${var.private_subnets}"]
   description = "Apiary DB Subnet Group"
@@ -12,11 +28,8 @@ resource "aws_db_subnet_group" "apiarydbsg" {
   tags = "${merge(map("Name","Apiary DB Subnet Group"),var.apiary_tags)}"
 }
 
-data "vault_generic_secret" "apiarydb_master_user" {
-  path = "${local.vault_path}/db_master_user"
-}
-
 resource "aws_security_group" "db_sg" {
+  count  = "${ var.external_database_host == "" ? 1 : 0 }"
   name   = "${local.instance_alias}-db"
   vpc_id = "${var.vpc_id}"
   tags   = "${var.apiary_tags}"
@@ -46,10 +59,12 @@ resource "aws_security_group" "db_sg" {
 }
 
 resource "random_id" "snapshot_id" {
+  count       = "${ var.external_database_host == "" ? 1 : 0 }"
   byte_length = 8
 }
 
 resource "aws_rds_cluster" "apiary_cluster" {
+  count                               = "${ var.external_database_host == "" ? 1 : 0 }"
   cluster_identifier                  = "${local.instance_alias}-cluster"
   database_name                       = "${var.apiary_database_name}"
   master_username                     = "${data.vault_generic_secret.apiarydb_master_user.data["username"]}"
@@ -69,7 +84,7 @@ resource "aws_rds_cluster" "apiary_cluster" {
 }
 
 resource "aws_rds_cluster_instance" "apiary_cluster_instance" {
-  count                = "${var.db_instance_count}"
+  count                = "${ var.external_database_host == "" ? var.db_instance_count : 0 }"
   identifier           = "${local.instance_alias}-instance-${count.index}"
   cluster_identifier   = "${aws_rds_cluster.apiary_cluster.id}"
   instance_class       = "${var.db_instance_class}"
@@ -83,6 +98,7 @@ resource "aws_rds_cluster_instance" "apiary_cluster_instance" {
 }
 
 resource "null_resource" "db_iam_auth" {
+  count      = "${ var.external_database_host == "" ? 1 : 0 }"
   depends_on = ["aws_rds_cluster_instance.apiary_cluster_instance"]
 
   provisioner "local-exec" {
@@ -90,28 +106,30 @@ resource "null_resource" "db_iam_auth" {
   }
 }
 
-resource "aws_route53_record" "apiarydb_alias" {
-  count   = "${local.enable_route53_records}"
-  zone_id = "${data.aws_route53_zone.apiary_zone.zone_id}"
-  name    = "${local.instance_alias}-metastore-db"
-  type    = "A"
+resource "null_resource" "mysql_rw_user" {
+  count      = "${ var.external_database_host == "" ? 1 : 0 }"
+  depends_on = ["aws_rds_cluster_instance.apiary_cluster_instance"]
 
-  alias {
-    name                   = "${aws_rds_cluster.apiary_cluster.endpoint}"
-    zone_id                = "${aws_rds_cluster.apiary_cluster.hosted_zone_id}"
-    evaluate_target_health = true
+  triggers {
+    username     = "${data.vault_generic_secret.hive_rwuser.data["username"]}"
+    password_md5 = "${md5(data.vault_generic_secret.hive_rwuser.data["password"])}"
+  }
+
+  provisioner "local-exec" {
+    command = "${path.module}/scripts/mysql-user.sh ${aws_rds_cluster.apiary_cluster.endpoint} ${aws_rds_cluster.apiary_cluster.master_username} '${aws_rds_cluster.apiary_cluster.master_password}' ALL ${var.apiary_database_name} ${data.vault_generic_secret.hive_rwuser.data["username"]} '${data.vault_generic_secret.hive_rwuser.data["password"]}'"
   }
 }
 
-resource "aws_route53_record" "apiarydb_ro_alias" {
-  count   = "${local.enable_route53_records}"
-  zone_id = "${data.aws_route53_zone.apiary_zone.zone_id}"
-  name    = "${local.instance_alias}-metastore-db-reader"
-  type    = "A"
+resource "null_resource" "mysql_ro_user" {
+  count      = "${ var.external_database_host == "" ? 1 : 0 }"
+  depends_on = ["aws_rds_cluster_instance.apiary_cluster_instance"]
 
-  alias {
-    name                   = "${aws_rds_cluster.apiary_cluster.reader_endpoint}"
-    zone_id                = "${aws_rds_cluster.apiary_cluster.hosted_zone_id}"
-    evaluate_target_health = true
+  triggers {
+    username     = "${data.vault_generic_secret.hive_rouser.data["username"]}"
+    password_md5 = "${md5(data.vault_generic_secret.hive_rouser.data["password"])}"
+  }
+
+  provisioner "local-exec" {
+    command = "${path.module}/scripts/mysql-user.sh ${aws_rds_cluster.apiary_cluster.endpoint} ${aws_rds_cluster.apiary_cluster.master_username} '${aws_rds_cluster.apiary_cluster.master_password}' SELECT ${var.apiary_database_name} ${data.vault_generic_secret.hive_rouser.data["username"]} '${data.vault_generic_secret.hive_rouser.data["password"]}'"
   }
 }